LAI2-7795: tabs for each layout item defined in app state
On app load, register a <Route> for each item defined in
'vars._layout' of the app state, unless it has a 'target'
property and it is treated as an external link.

For now, we just render all the children defined in the layout
item's 'content' as a separate iframe within the <Route>. In
the future, we will add the ability to rearrange using
drag-n-drop.