Feature(ui): generate and use OpenAPI client for state server
Adds the 'yarn run generate-client' command for generating
a TypeScript client for the Lightning CLI state server. Also
generates this client for the first time and uses it to
interact with the state server.

In the future, we can add a check to catch breaking changes in
the Lightning state API and the UI client.