Fix(ui) LAI2-8643: refetch app state every second (#73)
Adds a 'refetchInterval: 1000' to the 'useQuery()' call
which fetches the Lightning app state. This enables a more
reactive UI, for example when the app's layout changes and
there are different tabs to display.

This should eventually replaced with WebSockets once we figure
out how to get them working with FastAPI. bb827cc7978e3428ee99122e632ab6305b1c406a